Refactor log clean view into modal making log controller / model / view obsolete
